requirements.txt
transformers
peft
sentence-transformers
faiss-cpu
pandas
fastapi
uvicorn
torch
protobuf
sentencepiece
scikit-learn
tqdm
huggingface_hub

python-dotenv
